+++ src/index.js
@@ -14,6 +14,7 @@ import {
   getValidKeys,
   getValues,
   isArray,
+  isObject,
   isString
 } from './utils';
 
@@ -26,6 +27,10 @@ import {
   allSizeKeys,
   initialState
 } from './constants';
+
+const POSITION_PROP_DEFAULT = 'position';
+const RENDER_ON_RESIZE_DEFAULT = true;
+const SIZE_PROP_DEFAULT = 'size';
 
 let raf;
 
@@ -51,9 +56,16 @@ const setRaf = () => {
  *
  * @param {Component} OriginalComponent
  * @param {array<string>} keys
+ * @param {object} options={}
  * @returns {RemeasureComponent}
  */
-const getHigherOrderComponent = (OriginalComponent, keys) => {
+const getHigherOrderComponent = (OriginalComponent, keys, options = {}) => {
+  const {
+    positionProp = POSITION_PROP_DEFAULT,
+    renderOnResize = RENDER_ON_RESIZE_DEFAULT,
+    sizeProp = SIZE_PROP_DEFAULT
+  } = options;
+
   class RemeasureComponent extends Component {
     constructor(props) {
       super(props);
@@ -68,9 +80,11 @@ const getHigherOrderComponent = (OriginalComponent, keys) => {
 
       this.setValues(domElement);
 
-      onElementResize(domElement, () => {
-        this.setValues(domElement);
-      });
+      if (renderOnResize) {
+        onElementResize(domElement, () => {
+          this.setValues(domElement);
+        });
+      }
     }
 
     componentDidUpdate() {
@@ -129,7 +143,10 @@ const getHigherOrderComponent = (OriginalComponent, keys) => {
     };
 
     render() {
-      const values = getValues(keys, this.state);
+      const values = getValues(keys, this.state, {
+        positionProp,
+        sizeProp
+      });
 
       return (
         <OriginalComponent
@@ -148,9 +165,10 @@ const getHigherOrderComponent = (OriginalComponent, keys) => {
  * into OriginalComponent as an object under the prop name size and position
  *
  * @param {Component|array<string>} keys
+ * @param {object} options
  * @returns {RemeasureComponent}
  */
-const measure = (keys) => {
+const measure = (keys, options) => {
   if (isString(keys)) {
     switch (keys) {
       case 'size':
@@ -171,7 +189,13 @@ const measure = (keys) => {
     const validKeys = getValidKeys(keys, allKeys);
 
     return (OriginalComponent) => {
-      return getHigherOrderComponent(OriginalComponent, validKeys);
+      return getHigherOrderComponent(OriginalComponent, validKeys, options);
+    };
+  }
+
+  if (isObject(keys)) {
+    return (OriginalComponent) => {
+      return getHigherOrderComponent(OriginalComponent, allKeys, keys);
     };
   }